Committed changes in hw3
